--- app/domain/players/lib/game/daily_video.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module TopPop
-  module Game
-    # gameplay needed tools
-    class DailyVideos
-      def daily_videos
-        YoutubeInformation::Repository::Videos.all
-      end
-    end
-  end
-end
