+++ app/domain/players/lib/player_checker.rb
