+++ app/domain/videos/lib/generate_videos.rb
@@ -1,10 +1,12 @@
 # frozen_string_literal: true
 
-module YoutubeInformation
+module TopPop
   module Mixins
     # line credit calculation methods
-    module ScoreCalculator
-      
+    module  VideoGenerator
+      def G_daily_videos
+        TopPop::Repository::Videos#.random(5)
+      end
     end
   end
 end
