+++ spec/helpers/database_helper.rb
@@ -4,9 +4,9 @@
 module DatabaseHelper
   def self.wipe_database
     # Ignore foreign key constraints when wiping tables
-    YoutubeInformation::App.DB.run('PRAGMA foreign_keys = OFF')
-    YoutubeInformation::Database::VideoOrm.map(&:destroy)
-    YoutubeInformation::Database::PlayerOrm.map(&:destroy)
-    YoutubeInformation::App.DB.run('PRAGMA foreign_keys = ON')
+    TopPop::App.DB.run('PRAGMA foreign_keys = OFF')
+    TopPop::Database::VideoOrm.map(&:destroy)
+    TopPop::Database::PlayerOrm.map(&:destroy)
+    TopPop::App.DB.run('PRAGMA foreign_keys = ON')
   end
 end
