+++ spec/helpers/spec_helper.rb
@@ -17,5 +17,5 @@ require_app
 
 SEARCH_KEY_WORD = 'taylor%20swift%20offical'
 COUNT = 5
-YOUTUBE_TOKEN = YoutubeInformation::App.config.ACCESS_TOKEN
+YOUTUBE_TOKEN = TopPop::App.config.ACCESS_TOKEN
 CORRECT = YAML.safe_load(File.read('spec/fixtures/yt_results.yml'))
